+++ test/sparql_client/query/ask_test.exs
@@ -93,7 +93,8 @@ defmodule SPARQL.Client.Query.AskTest do
       end)
 
       assert SPARQL.Client.query(@example_query, @example_endpoint) ==
-               {:error, "unsupported result format: text/plain"}
+               {:error,
+                "SPARQL service responded with text/plain content which can't be interpreted. Try specifying one of the supported result formats with the :result_format option."}
     end
 
     test "when result_format set it's decoder is used", %{body: body} do