+++ test/sparql_client/query/select_test.exs
@@ -222,7 +222,8 @@ defmodule SPARQL.Client.Query.SelectTest do
       end)
 
       assert SPARQL.Client.query(@example_query, @example_endpoint) ==
-               {:error, "unsupported result format: text/html"}
+               {:error,
+                "SPARQL service responded with text/html content which can't be interpreted. Try specifying one of the supported result formats with the :result_format option."}
     end
 
     test "when result_format set it's decoder is used", %{body: body} do
@@ -274,7 +275,8 @@ defmodule SPARQL.Client.Query.SelectTest do
       assert SPARQL.Client.query(@example_query, @example_endpoint,
                headers: %{"Accept" => "text/plain"}
              ) ==
-               {:error, "unsupported result format: text/plain"}
+               {:error,
+                "SPARQL service responded with text/plain content which can't be interpreted. Try specifying one of the supported result formats with the :result_format option."}
     end
 
     test "with invalid format and result_format", %{body: body} do